raspberry-pi-voice-agent/requirements.txt
Porcupine
faster-whisper
paho-mqtt
sounddevice
numpy
gTTS
playsound
PyYAML